+++ CHANGES.rst
@@ -4,6 +4,9 @@ Release Notes
 In Development
 --------------
 
+- Fix ``Memory.cache(..., mmap_mode=..., verbose=...)`` to now uses its parameters
+  mmap_mode and verbose instead of using the values used at Memory creation time.
+
 - A non-positive value for ``pre_dispatch`` in ``Parallel`` now raises a
   ValueError.
   https://github.com/joblib/joblib/pull/1839